refactor: update app pages to use new atomic and feature-based components
